# [thrift-remodel] PoC new form for column index