… parquet/src/file/metadata/memory.rs
… parquet/src/file/metadata/mod.rs
… parquet/src/file/metadata/writer.rs
… parquet/src/file/page_index/column_index.rs
… parquet/src/file/page_index/index.rs
… parquet/src/file/page_index/index_reader.rs
… parquet/src/file/page_index/mod.rs
… parquet/src/file/serialized_reader.rs
… parquet/src/file/writer.rs
… parquet/tests/arrow_reader/io/mod.rs
… parquet/tests/encryption/encryption_util.rs