@etseidl 2025-08-20
This PR does not include the offset index changes from #8190. Running the metadata benchmark shows a substantial reduction in decode time.
```
old
open(page index)        time:   [1.7446 ms 1.7524 ms 1.7639 ms]

new with Index
open(page index)        time:   [698.00 µs 699.75 µs 701.66 µs]

new with ColumnIndexMetaData
open(page index)        time:   [422.47 µs 423.85 µs 425.18 µs]
```