@etseidl 2025-08-21
@alamb can you take a look when you have time? 🙏 Do you think the `ColumnIndexMetaData` here would work better for datafusion?