@etseidl 2025-08-26
I've got a good start on the write path now, so I'd like to merge this. But I would like a few more eyes on this on, given the radical change to the form of the column index.

And while we're breaking things, do we actually like the column and page indexes as `Vec<Vec<index>>`? I'm wondering if tacking them individually onto the `ColumnChunkMetaData` as `Option<index>` would be more intuitive? Or perhaps wrap them in `Arc` and store both places? Or if the primary use of this data is via the `StatisticsConverter` does this not really matter?

CC @jhorstmann @tustvold @XiangpengHao 

@adriangb you've also been in the metadata space, perhaps you have some opinions